+++ grpc-adapter/src/main/kotlin/jp/datachain/corda/ibc/grpc_adapter/HostService.kt
@@ -5,7 +5,7 @@ import ibc.lightclients.corda.v1.HostProto
 import ibc.lightclients.corda.v1.HostServiceGrpc
 import io.grpc.stub.StreamObserver
 import jp.datachain.corda.ibc.conversion.into
-import jp.datachain.corda.ibc.flows.IbcHostCreateFlow
+import jp.datachain.corda.ibc.flows.ics24.IbcHostCreateFlow
 import jp.datachain.corda.ibc.ics24.Host
 import net.corda.core.contracts.StateRef
 import net.corda.core.messaging.startFlow
