+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics3/IbcConnOpenConfirmFlow.kt
@@ -1,12 +1,13 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics3
 
 import co.paralleluniverse.fibers.Suspendable
-import ibc.core.channel.v1.Tx
+import ibc.core.connection.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics2.ClientState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleChanOpenTry
-import jp.datachain.corda.ibc.states.IbcChannel
+import jp.datachain.corda.ibc.ics26.HandleConnOpenConfirm
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
 import net.corda.core.contracts.StateRef
@@ -17,58 +18,45 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcChanOpenTryFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgChannelOpenTry
+class IbcConnOpenConfirmFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgConnectionOpenConfirm
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
-        // query host from vault
+        // query host state
         val host = serviceHub.vaultService.queryIbcHost(baseId)!!
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query conn from vault
-        val connId = Identifier(msg.channel.connectionHopsList.single())
-        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, connId)!!
-
-        // query client from vault
-        val clientId = Identifier(conn.state.data.end.clientId)
-        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, clientId)!!
-
-        // (optional) channel from vault
-        val chanOrNull = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, Identifier(msg.desiredChannelId))
+        // query conn state
+        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, Identifier(msg.connectionId))!!
+        // query client state
+        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, Identifier(conn.state.data.end.clientId))!!
 
         // create command and outputs
-        val command = HandleChanOpenTry(msg)
-        val inStates =
-                if(chanOrNull == null)
-                    setOf(host.state.data)
-                else
-                    setOf(host.state.data, chanOrNull.state.data)
-        val ctx = Context(inStates, setOf(client, conn).map{it.state.data})
+        val command = HandleConnOpenConfirm(msg)
+        val ctx = Context(setOf(conn.state.data), setOf(host, client).map{it.state.data})
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
         val notary = serviceHub.networkMapCache.notaryIdentities.single()
         val builder = TransactionBuilder(notary)
                 .addCommand(command, signers)
+                .addReferenceState(ReferencedStateAndRef(host))
                 .addReferenceState(ReferencedStateAndRef(client))
-                .addReferenceState(ReferencedStateAndRef(conn))
-                .addInputState(host)
-        chanOrNull?.let{builder.addInputState(it)}
+                .addInputState(conn)
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcChanOpenTryFlow::class)
-class IbcChanOpenTryResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcConnOpenConfirmFlow::class)
+class IbcConnOpenConfirmResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
