+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics3/IbcConnOpenInitFlow.kt
@@ -1,11 +1,13 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics3
 
 import co.paralleluniverse.fibers.Suspendable
-import ibc.core.channel.v1.Tx
+import ibc.core.connection.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
+import jp.datachain.corda.ibc.ics2.ClientState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleChanOpenInit
-import jp.datachain.corda.ibc.states.IbcConnection
+import jp.datachain.corda.ibc.ics26.HandleConnOpenInit
 import net.corda.core.contracts.ReferencedStateAndRef
 import net.corda.core.contracts.StateRef
 import net.corda.core.flows.*
@@ -15,45 +17,39 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcChanOpenInitFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgChannelOpenInit
+class IbcConnOpenInitFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgConnectionOpenInit
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
-        // query host from vault
         val host = serviceHub.vaultService.queryIbcHost(baseId)!!
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query connection from vault
-        val connId = Identifier(msg.channel.connectionHopsList.single())
-        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, connId)!!
+        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, Identifier(msg.clientId))!!
 
-        // create command and outputs
-        val command = HandleChanOpenInit(msg)
-        val ctx = Context(setOf(host.state.data), setOf(conn.state.data))
+        val command = HandleConnOpenInit(msg)
+        val ctx = Context(setOf(host.state.data), setOf(client.state.data))
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
-        // build tx
         val notary = serviceHub.networkMapCache.notaryIdentities.single()
         val builder = TransactionBuilder(notary)
-                .addCommand(command, signers)
+        builder.addCommand(command, ourIdentity.owningKey)
                 .addInputState(host)
-                .addReferenceState(ReferencedStateAndRef(conn))
+                .addReferenceState(ReferencedStateAndRef(client))
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcChanOpenInitFlow::class)
-class IbcChanOpenInitResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcConnOpenInitFlow::class)
+class IbcConnOpenInitResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
