+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics3/IbcConnOpenTryFlow.kt
@@ -1,12 +1,13 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics3
 
 import co.paralleluniverse.fibers.Suspendable
-import ibc.core.channel.v1.Tx
+import ibc.core.connection.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics2.ClientState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleChanOpenAck
-import jp.datachain.corda.ibc.states.IbcChannel
+import jp.datachain.corda.ibc.ics26.HandleConnOpenTry
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
 import net.corda.core.contracts.StateRef
@@ -17,54 +18,50 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcChanOpenAckFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgChannelOpenAck
+class IbcConnOpenTryFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgConnectionOpenTry
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
-        // query host from vault
+        // query host state
         val host = serviceHub.vaultService.queryIbcHost(baseId)!!
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query chan from vault
-        val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, Identifier(msg.channelId))!!
+        // query client state
+        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, Identifier(msg.clientId))!!
 
-        // query conn from vault
-        val connId = Identifier(chan.state.data.end.connectionHopsList.single())
-        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, connId)!!
+        // query conn state
+        val connOrNull = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, Identifier(msg.desiredConnectionId))
 
-        // query client from vault
-        val clientId = Identifier(conn.state.data.end.clientId)
-        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, clientId)!!
-
-        // create command and outputs
-        val command = HandleChanOpenAck(msg)
-        val ctx = Context(setOf(chan.state.data), setOf(host, client, conn).map{it.state.data})
+        val command = HandleConnOpenTry(msg)
+        val inStates =
+                if (connOrNull == null)
+                    setOf(host.state.data)
+                else
+                    setOf(host.state.data, connOrNull.state.data)
+        val ctx = Context(inStates, setOf(client.state.data))
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
-        // build tx
         val notary = serviceHub.networkMapCache.notaryIdentities.single()
         val builder = TransactionBuilder(notary)
                 .addCommand(command, signers)
-                .addReferenceState(ReferencedStateAndRef(host))
+                .addInputState(host)
                 .addReferenceState(ReferencedStateAndRef(client))
-                .addReferenceState(ReferencedStateAndRef(conn))
-                .addInputState(chan)
+        connOrNull?.let{builder.addInputState(it)}
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcChanOpenAckFlow::class)
-class IbcChanOpenAckResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcConnOpenTryFlow::class)
+class IbcConnOpenTryResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
