+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics4/IbcAcknowledgePacketFlow.kt
@@ -1,11 +1,14 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics4
 
 import co.paralleluniverse.fibers.Suspendable
 import ibc.core.channel.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcBank
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics2.ClientState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleChanOpenConfirm
+import jp.datachain.corda.ibc.ics26.HandlePacketAcknowledgement
 import jp.datachain.corda.ibc.states.IbcChannel
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
@@ -17,9 +20,9 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcChanOpenConfirmFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgChannelOpenConfirm
+class IbcAcknowledgePacketFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgAcknowledgement
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
@@ -28,8 +31,16 @@ class IbcChanOpenConfirmFlow(
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
+        // query bank if necessary
+        val bankOrNull =
+                if (msg.packet.destinationPort == "transfer")
+                    serviceHub.vaultService.queryIbcBank(baseId)!!
+                else
+                    null
+
         // query chan from vault
-        val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, Identifier(msg.channelId))!!
+        val chanId = Identifier(msg.packet.sourceChannel)
+        val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, chanId)!!
 
         // query conn from vault
         val connId = Identifier(chan.state.data.end.connectionHopsList.single())
@@ -40,11 +51,19 @@ class IbcChanOpenConfirmFlow(
         val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, clientId)!!
 
         // create command and outputs
-        val command = HandleChanOpenConfirm(msg)
-        val ctx = Context(setOf(chan.state.data), setOf(host, client, conn).map{it.state.data})
+        val command = HandlePacketAcknowledgement(msg)
+        val ctx = Context(
+                if (bankOrNull != null) {
+                    setOf(chan.state.data, bankOrNull.state.data)
+                } else {
+                    setOf(chan.state.data)
+                },
+                setOf(host, client, conn).map{it.state.data}
+        )
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
+        // build tx
         val notary = serviceHub.networkMapCache.notaryIdentities.single()
         val builder = TransactionBuilder(notary)
                 .addCommand(command, signers)
@@ -52,18 +71,18 @@ class IbcChanOpenConfirmFlow(
                 .addReferenceState(ReferencedStateAndRef(client))
                 .addReferenceState(ReferencedStateAndRef(conn))
                 .addInputState(chan)
+        bankOrNull?.let{builder.addInputState(it)}
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcChanOpenConfirmFlow::class)
-class IbcChanOpenConfirmResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcAcknowledgePacketFlow::class)
+class IbcAcknowledgePacketResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
