+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics4/IbcChanCloseConfirmFlow.kt
@@ -1,11 +1,14 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics4
 
 import co.paralleluniverse.fibers.Suspendable
-import ibc.core.connection.v1.Tx
+import ibc.core.channel.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics2.ClientState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleConnOpenAck
+import jp.datachain.corda.ibc.ics26.HandleChanCloseConfirm
+import jp.datachain.corda.ibc.states.IbcChannel
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
 import net.corda.core.contracts.StateRef
@@ -16,25 +19,31 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcConnOpenAckFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgConnectionOpenAck
+class IbcChanCloseConfirmFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgChannelCloseConfirm
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
-        // query host state
+        // query host from vault
         val host = serviceHub.vaultService.queryIbcHost(baseId)!!
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query conn state
-        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, Identifier(msg.connectionId))!!
-        // query client state
-        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, Identifier(conn.state.data.end.clientId))!!
+        // query chan from vault
+        val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, Identifier(msg.channelId))!!
+
+        // query conn from vault
+        val connId = Identifier(chan.state.data.end.connectionHopsList.single())
+        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, connId)!!
+
+        // query client from vault
+        val clientId = Identifier(conn.state.data.end.clientId)
+        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, clientId)!!
 
         // create command and outputs
-        val command = HandleConnOpenAck(msg)
-        val ctx = Context(setOf(conn.state.data), setOf(host.state.data, client.state.data))
+        val command = HandleChanCloseConfirm(msg)
+        val ctx = Context(setOf(chan.state.data), setOf(host, client, conn).map{it.state.data})
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
@@ -44,19 +53,19 @@ class IbcConnOpenAckFlow(
                 .addCommand(command, signers)
                 .addReferenceState(ReferencedStateAndRef(host))
                 .addReferenceState(ReferencedStateAndRef(client))
-                .addInputState(conn)
+                .addReferenceState(ReferencedStateAndRef(conn))
+                .addInputState(chan)
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcConnOpenAckFlow::class)
-class IbcConnOpenAckResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcChanCloseConfirmFlow::class)
+class IbcChanCloseConfirmResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
