+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics4/IbcChanCloseInitFlow.kt
@@ -1,11 +1,12 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics4
 
 import co.paralleluniverse.fibers.Suspendable
 import ibc.core.channel.v1.Tx
-import jp.datachain.corda.ibc.ics2.ClientState
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleChanCloseConfirm
+import jp.datachain.corda.ibc.ics26.HandleChanCloseInit
 import jp.datachain.corda.ibc.states.IbcChannel
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
@@ -17,9 +18,9 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcChanCloseConfirmFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgChannelCloseConfirm
+class IbcChanCloseInitFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgChannelCloseInit
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
@@ -28,20 +29,16 @@ class IbcChanCloseConfirmFlow(
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query chan from vault
+        // query channel from vault
         val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, Identifier(msg.channelId))!!
 
-        // query conn from vault
+        // query connection from vault
         val connId = Identifier(chan.state.data.end.connectionHopsList.single())
         val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, connId)!!
 
-        // query client from vault
-        val clientId = Identifier(conn.state.data.end.clientId)
-        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, clientId)!!
-
         // create command and outputs
-        val command = HandleChanCloseConfirm(msg)
-        val ctx = Context(setOf(chan.state.data), setOf(host, client, conn).map{it.state.data})
+        val command = HandleChanCloseInit(msg)
+        val ctx = Context(setOf(chan.state.data), setOf(host, conn).map{it.state.data})
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
@@ -50,7 +47,6 @@ class IbcChanCloseConfirmFlow(
         val builder = TransactionBuilder(notary)
                 .addCommand(command, signers)
                 .addReferenceState(ReferencedStateAndRef(host))
-                .addReferenceState(ReferencedStateAndRef(client))
                 .addReferenceState(ReferencedStateAndRef(conn))
                 .addInputState(chan)
         ctx.outStates.forEach{builder.addOutputState(it)}
@@ -58,13 +54,12 @@ class IbcChanCloseConfirmFlow(
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcChanCloseConfirmFlow::class)
-class IbcChanCloseConfirmResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcChanCloseInitFlow::class)
+class IbcChanCloseInitResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
