+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics4/IbcChanOpenAckFlow.kt
@@ -1,11 +1,13 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics4
 
 import co.paralleluniverse.fibers.Suspendable
 import ibc.core.channel.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics2.ClientState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandlePacketRecv
+import jp.datachain.corda.ibc.ics26.HandleChanOpenAck
 import jp.datachain.corda.ibc.states.IbcChannel
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
@@ -17,9 +19,9 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcRecvPacketFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgRecvPacket
+class IbcChanOpenAckFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgChannelOpenAck
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
@@ -28,16 +30,8 @@ class IbcRecvPacketFlow(
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query bank if necessary
-        val bankOrNull =
-                if (msg.packet.destinationPort == "transfer")
-                    serviceHub.vaultService.queryIbcBank(baseId)!!
-                else
-                    null
-
         // query chan from vault
-        val chanId = Identifier(msg.packet.destinationChannel)
-        val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, chanId)!!
+        val chan = serviceHub.vaultService.queryIbcState<IbcChannel>(baseId, Identifier(msg.channelId))!!
 
         // query conn from vault
         val connId = Identifier(chan.state.data.end.connectionHopsList.single())
@@ -48,15 +42,8 @@ class IbcRecvPacketFlow(
         val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, clientId)!!
 
         // create command and outputs
-        val command = HandlePacketRecv(msg)
-        val ctx = Context(
-                if (bankOrNull != null) {
-                    setOf(chan.state.data, bankOrNull.state.data)
-                } else {
-                    setOf(chan.state.data)
-                },
-                setOf(host, client, conn).map{it.state.data}
-        )
+        val command = HandleChanOpenAck(msg)
+        val ctx = Context(setOf(chan.state.data), setOf(host, client, conn).map{it.state.data})
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
@@ -68,19 +55,17 @@ class IbcRecvPacketFlow(
                 .addReferenceState(ReferencedStateAndRef(client))
                 .addReferenceState(ReferencedStateAndRef(conn))
                 .addInputState(chan)
-        bankOrNull?.let{builder.addInputState(it)}
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcRecvPacketFlow::class)
-class IbcRecvPacketResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcChanOpenAckFlow::class)
+class IbcChanOpenAckResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
