+++ workflows/src/main/kotlin/jp/datachain/corda/ibc/flows/ics4/IbcChanOpenInitFlow.kt
@@ -1,11 +1,12 @@
-package jp.datachain.corda.ibc.flows
+package jp.datachain.corda.ibc.flows.ics4
 
 import co.paralleluniverse.fibers.Suspendable
-import ibc.core.connection.v1.Tx
-import jp.datachain.corda.ibc.ics2.ClientState
+import ibc.core.channel.v1.Tx
+import jp.datachain.corda.ibc.flows.util.queryIbcHost
+import jp.datachain.corda.ibc.flows.util.queryIbcState
 import jp.datachain.corda.ibc.ics24.Identifier
 import jp.datachain.corda.ibc.ics26.Context
-import jp.datachain.corda.ibc.ics26.HandleConnOpenConfirm
+import jp.datachain.corda.ibc.ics26.HandleChanOpenInit
 import jp.datachain.corda.ibc.states.IbcConnection
 import net.corda.core.contracts.ReferencedStateAndRef
 import net.corda.core.contracts.StateRef
@@ -16,46 +17,44 @@ import net.corda.core.transactions.TransactionBuilder
 
 @StartableByRPC
 @InitiatingFlow
-class IbcConnOpenConfirmFlow(
-        val baseId: StateRef,
-        val msg: Tx.MsgConnectionOpenConfirm
+class IbcChanOpenInitFlow(
+        private val baseId: StateRef,
+        private val msg: Tx.MsgChannelOpenInit
 ) : FlowLogic<SignedTransaction>() {
     @Suspendable
     override fun call() : SignedTransaction {
-        // query host state
+        // query host from vault
         val host = serviceHub.vaultService.queryIbcHost(baseId)!!
         val participants = host.state.data.participants.map{it as Party}
         require(participants.contains(ourIdentity))
 
-        // query conn state
-        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, Identifier(msg.connectionId))!!
-        // query client state
-        val client = serviceHub.vaultService.queryIbcState<ClientState>(baseId, Identifier(conn.state.data.end.clientId))!!
+        // query connection from vault
+        val connId = Identifier(msg.channel.connectionHopsList.single())
+        val conn = serviceHub.vaultService.queryIbcState<IbcConnection>(baseId, connId)!!
 
         // create command and outputs
-        val command = HandleConnOpenConfirm(msg)
-        val ctx = Context(setOf(conn.state.data), setOf(host, client).map{it.state.data})
+        val command = HandleChanOpenInit(msg)
+        val ctx = Context(setOf(host.state.data), setOf(conn.state.data))
         val signers = listOf(ourIdentity.owningKey)
         command.execute(ctx, signers)
 
+        // build tx
         val notary = serviceHub.networkMapCache.notaryIdentities.single()
         val builder = TransactionBuilder(notary)
                 .addCommand(command, signers)
-                .addReferenceState(ReferencedStateAndRef(host))
-                .addReferenceState(ReferencedStateAndRef(client))
-                .addInputState(conn)
+                .addInputState(host)
+                .addReferenceState(ReferencedStateAndRef(conn))
         ctx.outStates.forEach{builder.addOutputState(it)}
 
         val tx = serviceHub.signInitialTransaction(builder)
 
         val sessions = (participants - ourIdentity).map{initiateFlow(it)}
-        val stx = subFlow(FinalityFlow(tx, sessions))
-        return stx
+        return subFlow(FinalityFlow(tx, sessions))
     }
 }
 
-@InitiatedBy(IbcConnOpenConfirmFlow::class)
-class IbcConnOpenConfirmResponderFlow(val counterPartySession: FlowSession) : FlowLogic<Unit>() {
+@InitiatedBy(IbcChanOpenInitFlow::class)
+class IbcChanOpenInitResponderFlow(private val counterPartySession: FlowSession) : FlowLogic<Unit>() {
     @Suspendable
     override fun call() {
         val stx = subFlow(ReceiveFinalityFlow(counterPartySession))
